Python and C++ API work
Former-commit-id: 69e3d39b9f86981b20002bc942d7c60af1cd4f13 [formerly 839b45f55fb056da5c4d37aa0585d7dc2cf6d0e6]
Former-commit-id: d59b6d316a99af111a70b5270db104267020b1c8
Former-commit-id: 8715034961b70061e6aa72e4775ebe654c9ea52b

+++ Halite/Classes/Networking.h
@@ -36,7 +36,7 @@ static std::string serializeMap(hlt::Map & map)
             }
             else
             {
-                oss << counter << " " << currentOwner << " ";
+                oss << (unsigned short)counter << " " << (unsigned short)currentOwner << " ";
                 counter = 1;
                 currentOwner = map.contents[a][b].owner;
             }
@@ -50,7 +50,7 @@ static std::string serializeMap(hlt::Map & map)
     {
         for (int b = 0; b < map.contents[a].size(); ++b)
         {
-            oss << map.contents[a][b].strength << " ";
+            oss << (unsigned short)map.contents[a][b].strength << " ";
         }
     }
